Fixed pre-existing tests

+++ internal/innosat/source.go
@@ -23,7 +23,7 @@ func (headerType SourcePacketHeaderType) String() string {
 	case TC:
 		return "TC"
 	default:
-		return fmt.Sprintf("Unknown %v", headerType)
+		return fmt.Sprintf("Unknown %v", uint(headerType))
 	}
 }
 
@@ -52,7 +52,7 @@ func (continuationFlag SourcePackageContinuationFlagType) String() string {
 	case SPStandalone:
 		return "Standalone"
 	default:
-		return fmt.Sprintf("Unknown %v", continuationFlag)
+		return fmt.Sprintf("Unknown %v", uint(continuationFlag))
 	}
 }
